% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes.R
\name{convert_gene}
\alias{convert_gene}
\title{Convert a mouse gene to a human gene}
\usage{
convert_gene(gene, genes_df, ogene, ngene)
}
\arguments{
\item{gene}{Name of the mouse gene}

\item{genes_df}{Dataframe containing the conversion between the new and old gene names}
}
\value{
Name of the human gene, NaN if no match
}
\description{
This function converts the given mouse gene to the matching human gene. If
there is no matching mouse gene, then we return NaN.
}
