% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genes.R
\name{convert_genes}
\alias{convert_genes}
\title{Adds a column to the df with the human gene name corresponding to the mouse gene}
\usage{
convert_genes(df, conversion_file, old_gene_col, new_gene_col)
}
\arguments{
\item{conversion_file}{.csv file containing equivalence b/w human and mouse
genes in two columns (HumanSymbol & MouseSymbol)}

\item{old_gene_col}{Name of column containing the old gene name}

\item{new_gene_col}{Name of column containing the new gene name}

\item{gene}{Name of the mouse gene}
}
\value{
Dataframe with a new column containing the new gene names
}
\description{
This function adds a column \code{HumanSymbol} to the given df containing the
human gene name for the gene present in the column \code{MouseSymbol}. If no matching gene
is found, it adds a NaN value to the column. Uses a conversion file between mouse
and human gene symbols (parameter \code{conversion_file}).
}
