% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-to-mouse.R
\name{convert_genes_human_to_mouse}
\alias{convert_genes_human_to_mouse}
\title{Adds a column to the df with the mouse gene name corresponding to the human gene}
\usage{
convert_genes_human_to_mouse(df, conversion_file)
}
\arguments{
\item{df}{Dataframe containing the \code{HumanSymbol} column}
}
\value{
Dataframe identical to \code{df}, with an added column called \code{MouseSymbol}
}
\description{
This function adds a column \code{MouseSymbol} to the given df containing the
mouse gene name for the gene present in the column \code{HumanSymbol}. If no matching gene
is found, it adds a NaN value to the column. Uses a conversion file between mouse
and human gene symbols (parameter \code{conversion_file}).
}
