% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse-to-human.R
\name{convert_genes_mouse_to_human}
\alias{convert_genes_mouse_to_human}
\title{Adds a column to the df with the human gene name corresponding to the mouse gene}
\usage{
convert_genes_mouse_to_human(df, conversion_file)
}
\arguments{
\item{df}{Dataframe containing the \code{MouseSymbol} column}
}
\value{
Dataframe identical to \code{df}, with an added column called \code{HumanSymbol}
}
\description{
This function adds a column \code{HumanSymbol} to the given df containing the
human gene name for the gene present in the column \code{MouseSymbol}. If no matching gene
is found, it adds a NaN value to the column. Uses a conversion file between mouse
and human gene symbols (parameter \code{conversion_file}).
}
